requirements.txt
xai-sdk
openai
aiohttp
websockets
polygon-api-client
streamlit
pandas
numpy
networkx
chromadb
sentence-transformers
sec-edgar-downloader
plyer
requests
scipy
statsmodels
pandas_datareader
yfinance
tweepy
langgraph
langchain
langchain-community
playwright
streamlit-aggrid
plotly
python-dotenv
langchain-openai
beautifulsoup4
fpdf
streamlit-mic-recorder
gTTS